+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/ReverseGeocodingSearchIntegrationTest.kt
@@ -13,11 +13,11 @@ import com.mapbox.search.common.FixedPointLocationEngine
 import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.SearchRequestException
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
-import com.mapbox.search.metadata.OpenHours
-import com.mapbox.search.metadata.OpenPeriod
-import com.mapbox.search.metadata.ParkingData
-import com.mapbox.search.metadata.WeekDay
-import com.mapbox.search.metadata.WeekTimestamp
+import com.mapbox.search.common.metadata.OpenHours
+import com.mapbox.search.common.metadata.OpenPeriod
+import com.mapbox.search.common.metadata.ParkingData
+import com.mapbox.search.common.metadata.WeekDay
+import com.mapbox.search.common.metadata.WeekTimestamp
 import com.mapbox.search.record.FavoritesDataProvider
 import com.mapbox.search.record.HistoryDataProvider
 import com.mapbox.search.result.ResultAccuracy