+++ MapboxSearch/sdk/src/androidTest/java/com/mapbox/search/SearchEngineIntegrationTest.kt
@@ -17,8 +17,8 @@ import com.mapbox.search.common.RoutablePoint
 import com.mapbox.search.common.SearchRequestException
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
 import com.mapbox.search.common.equalsTo
-import com.mapbox.search.metadata.OpenHours
-import com.mapbox.search.metadata.ParkingData
+import com.mapbox.search.common.metadata.OpenHours
+import com.mapbox.search.common.metadata.ParkingData
 import com.mapbox.search.record.FavoritesDataProvider
 import com.mapbox.search.record.HistoryDataProvider
 import com.mapbox.search.record.HistoryRecord