+++ MapboxSearch/sdk/src/main/java/com/mapbox/search/utils/serialization/OpenModeDAO.kt
@@ -2,7 +2,7 @@ package com.mapbox.search.utils.serialization
 
 import com.google.gson.annotations.SerializedName
 import com.mapbox.search.base.utils.printableName
-import com.mapbox.search.metadata.OpenHours
+import com.mapbox.search.common.metadata.OpenHours
 
 internal enum class OpenModeDAO {